Type of FileDescriptor.size should be Long #PL-3867

+++ modules/core/db/init/hsql/create-db.sql
@@ -51,7 +51,7 @@ create table SYS_FILE (
     --
     NAME varchar(500),
     EXT varchar(20),
-    FILE_SIZE integer,
+    FILE_SIZE bigint,
     CREATE_DATE timestamp,
     --
     primary key (ID)
